lots of updates

+++ src/accountbook/account.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 from datetime import datetime
-from pathlib import Path
 from collections import defaultdict
 import json
 
@@ -15,16 +14,16 @@ class Account:
     owner: str
     bank: str
     vault: str
-    default_balance: float = field(repr=False)
     id: str = field(init=False)
-    _balance: float = field(init=False, repr=False)
-    balance: float
     description: str = None
+    default_balance: float = field(default=None, repr=False)
+    init: bool = field(default=False, repr=False)
+    _balance: float = field(init=False, repr=False)
 
     def __post_init__(self) -> None:
         self.id: str = f"{self.owner}-{self.bank}-{self.vault}"
         self.file_path = ACCOUNTS / f"{self.id}.json"
-        if not self.file_path.exists():
+        if self.init and not self.file_path.exists():
             self.initiate_file()
 
     @property
@@ -33,33 +32,48 @@ class Account:
 
     @balance.setter
     def balance(self, new_value: float):
-        if not isinstance(new_value, (float, int)):
-            raise TypeError("Balance is numeric.")
+        assert isinstance(new_value, (float, int)), TypeError("Balance is numeric.")
         self._balance = round(float(new_value), 2)
-        self.last_balance_update = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-
-        file_path = ACCOUNTS / f"{self.owner}-{self.bank}-{self.vault}.json"
-        if file_path.exists():
-            with open(file_path) as file:
+        if self.file_path and self.file_path.exists():
+            with open(self.file_path) as file:
                 data = json.load(file)
-            with open(file_path, "w") as file:
-                data["profile"]["balance"] = self._balance
-                data["profile"]["last_balance_update"] = self.last_balance_update
-                print(f"Balance set to {self._balance} on {self.last_balance_update}")
+            with open(self.file_path, "w") as file:
+                data["balance"] = self._balance
+                data["latest_balance_update"] = datetime.now().strftime(
+                    "%Y-%m-%d %H:%M:%S"
+                )
+                print(f"Balance set to {self._balance:.2f}")
                 json.dump(data, file, indent=4)
 
     @staticmethod
-    def from_file(file_path: Path | str) -> Account:
-        with open(file_path) as f:
-            attributes = json.load(f)["profile"]
-        return Account(**attributes)
+    def from_id(account_id: str, **kwargs) -> Account:
+        owner, bank, vault = account_id.split("-")
+        return Account(owner=owner, bank=bank, vault=vault, **kwargs)
 
     def initiate_file(self) -> None:
-        profile_attributes = config_account["profile_attributes"]
-        profile = {attr: getattr(self, attr) for attr in profile_attributes}
+        ACCOUNTS.mkdir(parents=True, exist_ok=True)
+        assert self.default_balance is not None, ValueError(
+            "Default balance must be set."
+        )
+        profile = {
+            attr: getattr(self, attr) for attr in config_account["profile_attributes"]
+        }
+        data = dict(balance=self.default_balance, profile=profile)
+        data["latest_balance_update"] = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+        data["latest_profile_update"] = data["latest_balance_update"]
+
         with open(self.file_path, "w") as file:
-            json.dump(dict(profile=profile), file, indent=4)
-        print(f"Created a new file {self.file_path}")
+            json.dump(data, file, indent=4)
+            print(f"Created a new file {self.file_path}")
+
+    def update_profile(self, **kwargs) -> None:
+        with open(self.file_path) as file:
+            data = json.load(file)
+        data["profile"].update(kwargs)
+        data["latest_profile_update"] = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+        with open(self.file_path, "w") as file:
+            json.dump(data, file, indent=4)
+            print("Updated profile.")
 
     def load(self):
         with open(self.file_path) as file:
@@ -74,8 +88,9 @@ class Account:
             data = json.load(file, object_hook=fn_defaultdict_list)
 
         snapshot = {
-            "date": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
-            "balance": self.balance,
+            "timestamp": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
+            "profile": data["profile"],
+            "balance": data["balance"],
         }
         if label:
             snapshot["label"] = label
@@ -86,17 +101,17 @@ class Account:
 
 
 def main():
-    a = Account(*AccountEnum.TEST.value, balance=100)
+    a = Account.from_id(AccountEnum.TEST.value)
+    a.file_path.unlink(missing_ok=True)
+    a.default_balance = 100
+    a.description = "Test account"
+    a.initiate_file()
     print(f"{a=}")
     a.balance = 150
-    print(f"{a=}")
     a.take_snapshot()
+    a.update_profile(description="Test account for testing")
     a.balance = 200
-    a.take_snapshot(label="Xmas")
-
-    b = Account.from_file(a.file_path)
-    print(f"{b=}")
-    b.take_snapshot()
+    a.take_snapshot(label="Cheese!")
 
 
 if __name__ == "__main__":
